+++ package.json
@@ -24,7 +24,7 @@
     "expo-firebase-recaptcha": "~2.1.0",
     "expo-image-picker": "~11.0.3",
     "expo-status-bar": "~1.1.0",
-    "firebase": "^9.4.1",
+    "firebase": "^9.6.1",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "^0.66.4",
