+++ src/config/firebaseConfig.js
@@ -1,5 +1,5 @@
 import { initializeApp } from "firebase/app";
-import { getAuth } from "fiebase/auth";
+import { getAuth } from "firebase/auth";
 import { getFirestore } from "firebase/firestore";
 
 const firebaseConfig = {
