+++ src/styles/MainPage.js
