+++ src/styles/UploadCancelPopup.js
